Add files via upload
This program calculates the tuition a student owes based on factors such as their NYC residency status and credits amount.  I used classes and class inheritance in order to create objects within the main code. I created 3 classes: a superclass named Person, a subclass named Student, and a separate class named Loan.  After the program determines the amount in which the student owes, it allows the student to enter the amount they desire to pay. Based on conditional factors if the amount that the student enters is equivalent to how much they owe the program ends, if not the program continues and has the user pay off the remaining balance with a loan object.